requirements.txt
flask
mongoengine
python-dotenv
pyjwt
flask-cors
vosk
pyttsx3
wave
speechrecognition
google-generativeai
